텍스트 모델 나무성장 연동

+++ src/main/java/com/treetment/backend/Emotiontree/EmotiontreeService.java
@@ -1,11 +1,16 @@
 package com.treetment.backend.Emotiontree;
 
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.CopyObjectRequest;
+import com.treetment.backend.auth.entity.User;
+import com.treetment.backend.auth.repository.UserRepository;
 import com.treetment.backend.entity.EmotionTree;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
+import org.springframework.beans.factory.annotation.Value;
+import java.util.Optional;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -15,7 +20,7 @@ public class EmotiontreeService {
     private final EmotiontreeRepository emotiontreeRepository;
 
     @Transactional(readOnly = true)
-    public List<EmotiontreeDTO> getTreesByUserId(Long userId)
+    public List<EmotiontreeDTO> getTreesByUserId(Integer userId)
     {
         List<EmotionTree> emotionTrees = emotiontreeRepository.findByUserIdWithUser(userId);
         return emotionTrees.stream()
